+++ barcode_blastn/controllers/blastdb_controller.py
@@ -1,12 +1,12 @@
 from datetime import datetime, timezone
 import os
 import shutil
-from typing import Any, Dict, List, Optional, Set, Tuple
+from typing import Any, Dict, List, Optional, Set, Tuple, Union
 from barcode_blastn.file_paths import get_data_fishdb_path, get_data_library_path, get_ncbi_folder
 from barcode_blastn.helper.parse_gb import GenBankConnectionError, InsufficientAccessionData, retrieve_gb, save_taxonomy
 from django.contrib.auth.models import User
 from django.db import transaction
-from barcode_blastn.models import Annotation, BlastDb, Hit, Library, NuccoreSequence
+from barcode_blastn.models import Annotation, BlastDb, CustomSequence, Hit, Library, NuccoreSequence
 from barcode_blastn.serializers import NuccoreSequenceSerializer 
 from django.db.models import QuerySet, Value, Q
 from django.db.models.functions import Length, Replace, Upper
@@ -178,8 +178,11 @@ def create_blastdb(additional_accessions: List[str], user: User, base: Optional[
     
     accessions_to_add = additional_accessions
     seqs: Optional[QuerySet[NuccoreSequence]]
+
+    # Clone the accessions from the base database to the new db
     if base is not None:
-        seqs = NuccoreSequence.objects.filter(owner_database=base)
+        seqs = NuccoreSequence.objects.filter(owner_database=base, 
+            data_source=NuccoreSequence.SequenceSource.GENBANK)
         # Add the base database's accessions to the list of accessions to retrieve
         accessions_to_add.extend([s.version for s in seqs])
     else:
@@ -469,7 +472,16 @@ def filter_sequences_in_database(database: BlastDb, user: User, min_length: int 
     save_blastdb(database, user=user, perform_lock=False)
     return deleted_objects
 
-def add_sequences_to_database(database: BlastDb, user: User, desired_numbers: List[str] = [], annotations_to_transfer: Dict[str, Any] = {}, search_term: Optional[str] = None, min_length: int = -1, max_length: int = -1, max_ambiguous_bases = -1, blacklist: List[str] = [], require_taxonomy: bool = False) -> List[NuccoreSequence]:
+def add_sequences_to_database(database: BlastDb, 
+                                user: User, 
+                                desired_numbers: List[str] = [], 
+                                annotations_to_transfer: Dict[str, Any] = {}, 
+                                search_term: Optional[str] = None, 
+                                min_length: int = -1, 
+                                max_length: int = -1, 
+                                max_ambiguous_bases = -1, 
+                                blacklist: List[str] = [], 
+                                require_taxonomy: bool = False) -> List[NuccoreSequence]:
     '''
     Add a list of accession numbers to an existing database by bulk creating and return the resulting list of NuccoreSequences, and return the saved objects.
     
@@ -572,6 +584,123 @@ def add_sequences_to_database(database: BlastDb, user: User, desired_numbers: Li
     save_blastdb(database, user=user, perform_lock=False)
     return to_create
 
+def add_custom_sequences_to_database(database: BlastDb,
+                                user: User, 
+                                custom_sequences: QuerySet[CustomSequence],
+                                annotations_to_transfer: Dict[str, Any] = {},
+                                min_length: int = -1, 
+                                max_length: int = -1, 
+                                max_ambiguous_bases = -1, 
+                                blacklist: List[str] = [], 
+                                require_taxonomy: bool = False) -> List[CustomSequence]:
+    """
+    Clone custom sequences into the new database.
+    """
+
+    # Apply function to filter the existing custom_sequences
+    to_clone: List[CustomSequence] = []
+    for seq in custom_sequences:
+        seq_len = seq.sequence_length()
+        if min_length > -1 and seq_len < min_length:
+            continue
+        if max_length > -1 and seq_len > max_length:
+            continue
+        
+        # filter by ambiguous bases
+        n = len([base for base in seq.dna_sequence if base == 'N'])
+        if max_ambiguous_bases > -1 and n > max_ambiguous_bases:
+            continue
+
+        # require taxonomy
+        if require_taxonomy:
+            taxa = ['superkingdom', 'kingdom', 'phylum', 'class', 'order', 'family', 'genus', 'species']
+            if any([(getattr(seq, f'taxon_{t}', None) is None) for t in taxa]):
+                continue
+        
+        # check if in blacklist
+        if seq.accession_number in blacklist or seq.version in blacklist:
+            continue
+
+        to_clone.append(seq)
+
+    # Copy over annotations
+    annotations: List[List[Dict[str, Any]]] = []
+    for seq in to_clone:
+        annot: List[Annotation]
+        annot = list(seq.annotations.all()) # type: ignore
+        annot_data: List[Dict[str, Any]] = [{
+            'annotation_type': getattr(a, 'annotation_type', Annotation.AnnotationType.OTHER),
+            'comment': getattr(a, 'comment', '')
+        } for a in annot]
+        annot_data.extend(annotations_to_transfer.get(seq.version, []))
+        annotations.append(annot_data)
+            
+    # Set primary key to None to force duplicate, then
+    # save manually
+    seq: CustomSequence
+    for i, seq in enumerate(to_clone):
+        seq.pk = None
+        # Pass info to save signal for saving
+        setattr(seq, 'create_annotations', annotations[i])
+        setattr(seq, 'annotation_user', user)
+
+        # Manually call save on each instance, instead of using .bulk_create,
+        # to ensure post_save signal is sent.
+        seq.save()
+
+    # Add the initial accessions and search terms to be stored in the 
+    # historical log of the database
+    log_added_sequences(instances=to_clone, search_term='', database=database)
+    log_filter_options(database, min_length=min_length, max_length=max_length, max_ambiguous_bases=max_ambiguous_bases, \
+        blacklist=blacklist, require_taxonomy=require_taxonomy)
+    save_blastdb(database, user=user, perform_lock=False)
+        
+    return to_clone
+
+def save_custom_sequence(seqs: List[CustomSequence], user: User, raise_errors: bool = True):
+    '''
+    Save a custom sequence, where the user specifies the value of every value
+    in the model.
+
+    Raises:
+        
+    '''
+    for obj in seqs:
+
+        value_error: Optional[ValueError] = None
+        if not obj.accession_number or len(obj.accession_number) == 0:
+            value_error = ValueError('Missing accession numbers')
+        elif not obj.owner_database:
+            value_error = ValueError()
+        elif obj.owner_database.locked:
+            value_error = ValueError('Database locked')
+        if raise_errors and not value_error is None:
+            raise value_error
+        else:
+            print(f'WARN: Suppressed error {value_error}')
+
+        # Ensure that the obj source is marked as import
+        obj.data_source = NuccoreSequence.SequenceSource.IMPORT
+        
+        # Check that the accession does not yet exist in the database.
+        try:
+            existing = NuccoreSequence.objects.filter(owner_database=obj.owner_database).exclude(id=obj.id).get(accession_number=obj.accession_number)
+        except NuccoreSequence.DoesNotExist:
+            pass
+        else:
+            # raise error if duplicate accession already exists
+            error = AccessionsAlreadyExist([existing.accession_number])
+            if raise_errors:
+                raise error 
+            else:
+                print(f'WARN: Suppressed error {error}')
+                
+    # Save separately on another pass, to ensure that all seqs pass checks above
+    # first
+    for obj in seqs:
+        obj.save()
+
+
 def save_sequence(obj: NuccoreSequence, user: User, commit: bool = False, raise_if_missing: bool = False, raise_errors: bool = True):
     '''
     Save an accession by fetching data from GenBank again and populating the instance. Used for both saving and updating an accession.
@@ -601,6 +730,9 @@ def save_sequence(obj: NuccoreSequence, user: User, commit: bool = False, raise_
         raise value_error
     else:
         print(f'WARN: Suppressed error {value_error}')
+
+    # Ensure that the source is GenBank
+    obj.data_source = NuccoreSequence.SequenceSource.GENBANK
 
     accession_number = obj.accession_number
 
@@ -720,7 +852,7 @@ def log_filter_options(database: BlastDb, min_length: int = -1, max_length: int 
     print(database.id, min_length)
     setattr(database, 'filter_options', '\n'.join(options))
 
-def log_added_sequences(instances: List[NuccoreSequence], search_term: Optional[str], database: BlastDb) -> None:
+def log_added_sequences(instances: Union[List[NuccoreSequence], List[CustomSequence]], search_term: Optional[str], database: BlastDb) -> None:
     '''
     Modify the `._change_reason`, `.added`, and `.search_terms` of database to reflect the addition of 
     sequences in instances. If the object database already has non-zero-length `._change_reason`,